feat: Introduce file upload question type, update Docker configuration to use Bun, and add Keycloak authentication types.
