+++ docker-compose.yml
@@ -27,8 +27,8 @@ services:
             MINIO_ROOT_PASSWORD: ${MINIO_ROOT_PASSWORD}
         command: server /data --console-address ":9001"
         ports:
-            - "${MINIO_API_PORT:-9000}:9000"
-            - "${MINIO_CONSOLE_PORT:-9001}:9001"
+            - "${S3_API_PORT:-9000}:9000"
+            - "${S3_CONSOLE_PORT:-9001}:9001"
         volumes:
             - minio_data:/data
         networks:
