+++ env.app.example
@@ -2,7 +2,6 @@
 VERSION="0.0.0"
 
 # Authentication
-AUTH_SECRET=AUTH_SECRET
 NEXT_REDIRECT=http://localhost:3000
 
 # Logging
@@ -34,7 +33,7 @@ NEXT_PUBLIC_POSTHOG_ASSETS_HOST=https://us-assets.i.posthog.com
 NEXT_PUBLIC_POSTHOG_ENABLE="false"
 
 # NextAuth (Legacy/Compat)
-NEXTAUTH_SECRET=AUTH_SECRET
+NEXTAUTH_SECRET=NEXTAUTH_SECRET
 
 # Keycloak Admin Credentials (for user sync)
 KEYCLOAK_ADMIN_USERNAME=admin
@@ -48,7 +47,7 @@ KEYCLOAK_REALM=evalify
 
 # Evaluation Service
 EVALUATION_SERVICE_URL=http://localhost:4000
-EVALUATION_SERVICE_API_KEY=evalify_api_key
+API_KEY=evalify_api_key
 
 # External APIs
 PROTECTED="true"  # Set to false to disable API protection only in development. NOT RECOMMENDED FOR PRODUCTION
