+++ env.docker-compose.example
@@ -12,8 +12,8 @@ POSTGRES_PORT=5432
 # MinIO
 MINIO_ROOT_USER=minioadmin
 MINIO_ROOT_PASSWORD=minioadmin123
-MINIO_API_PORT=9000
-MINIO_CONSOLE_PORT=9001
+S3_API_PORT=9000
+S3_CONSOLE_PORT=9001
 
 # Redis / Valkey
 REDIS_PORT=6379
