+++ package.json
@@ -6,6 +6,11 @@
         "dev": "next dev --turbopack",
         "build": "next build --turbopack",
         "start": "next start",
+        "dev:bun": "bun --bun next dev",
+        "build:bun": "bun --bun next build",
+        "start:bun": "bun --bun next start",
+        "start:server": "node .next/standalone/server.js",
+        "start:server:bun": "bun .next/standalone/server.js",
         "lint": "eslint",
         "lint:fix": "eslint --fix",
         "format": "prettier --write .",
