+++ src/components/providers/providers.tsx
@@ -17,7 +17,7 @@ export function Providers({ children }: ProvidersProps) {
     return (
         <NextThemesProvider
             attribute="class"
-            defaultTheme="system"
+            defaultTheme="dark"
             enableSystem
             disableTransitionOnChange
         >
